requirements.txt
python-telegram-bot==21.6
google-genai>=1.0.0
yfinance==0.2.50
APScheduler==3.10.4
requests==2.32.3
beautifulsoup4==4.12.3
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
feedparser==6.0.11
tabulate==0.9.0
aiohttp==3.10.5
lxml==5.3.0